bugfix(?) and performance improvement
Bugfix: the central directory header's local header offset field was theoretically overflowing when it exceeded 4GB instead of being written to the ZIP64 extra field, which it is now. However, the program I use to test the zip wasn't flagging it as corrupt before so, although it should be now, I can't be sure if it's correct now.

Performance Improvement: now compresses and writes the file's data first and only then the header all at once instead of header, file data, and header corrections.